PurvaPatil0911 Create Attendance Management System .java
Task 3 - Attendance Management System in Java

Develop an Attendance Management System using Java, following the MVC architecture and utilizing Maven as a build tool. MySQL will be used as the backend database, and the application will be deployed on a Tomcat server for live access. This system caters to universities, schools, and organizations for tracking student or employee attendance. 

Key Features : 
- Separate user roles: Admin, Student, Faculty. 
- Admin privileges for adding, updating, and deleting user records. 
- User-friendly interface for marking and viewing attendance. 
- Database integration for storing attendance data